# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Town Backend",
    "description": "二次元 AI 小镇陪伴智能体 - World Engine + LLM",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/characters/{character_id}/messages": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Get Character Messages",
        "description": "获取角色的所有消息历史（跨会话）\n\nArgs:\n    character_id: 角色 ID\n    limit: 返回数量上限\n\nReturns:\n    消息列表（按时间正序）",
        "operationId": "get_character_messages_api_v1_characters__character_id__messages_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Character Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Character Messages Api V1 Characters  Character Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/messages/send": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send Message",
        "description": "发送消息给角色并获取回复\n\nArgs:\n    character_id: 角色 UUID\n    user_id: 用户标识\n    platform: 来源平台（web/qq/lark/internal）\n    content: 用户消息内容\n\nReturns:\n    角色回复内容与元数据（token/cost/conversation_id）",
        "operationId": "send_message_api_v1_messages_send_post",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "web",
              "title": "Platform"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Content"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_message_api_v1_messages_send_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/history": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Get Message History",
        "description": "获取会话消息历史（支持游标分页）\n\nArgs:\n    conversation_id: 会话 UUID\n    limit: 返回数量上限（默认 50）\n    before: 游标时间（ISO 8601），仅返回该时间点之前的消息\n\nReturns:\n    消息列表（按时间倒序）",
        "operationId": "get_message_history_api_v1_messages_history_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageHistoryListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List Conversations",
        "description": "查询会话列表（仅返回当前用户的会话）\n\nArgs:\n    current_user: 当前认证用户（JWT/API Key）\n    character_id: 可选，按角色过滤\n    user_id: 可选，按用户过滤；与当前用户不一致时拒绝\n    limit: 返回数量上限\n\nReturns:\n    会话列表（按 last_message_at 倒序）",
        "operationId": "list_conversations_api_v1_conversations_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Character Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationsListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/stats": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Get Message Stats",
        "description": "获取消息统计（token/cost 累计，供成本监控）\n\nArgs:\n    character_id: 可选，按角色过滤\n    start_date: 可选，起始日期（ISO 8601）\n    end_date: 可选，结束日期（ISO 8601）\n\nReturns:\n    累计消息数、token 数与 cost（USD），含按角色/按日期分组",
        "operationId": "get_message_stats_api_v1_messages_stats_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Character Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageStatsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/onebot/messages": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Onebot Messages",
        "description": "获取 QQ 消息记录（用于 QQ 消息监控）\n\n查询 platform=qq 的会话中的最近消息，包含发送者和内容。\n\nReturns:\n    消息列表（按时间倒序）",
        "operationId": "get_onebot_messages_api_v1_admin_onebot_messages_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnebotMessagesListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_send_message_api_v1_messages_send_post": {
        "properties": {
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "user_id"
        ],
        "title": "Body_send_message_api_v1_messages_send_post"
      },
...
      "ConversationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "last_message_at": {
            "type": "string",
            "title": "Last Message At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "character_id",
          "user_id",
          "platform",
          "last_message_at"
        ],
        "title": "ConversationOut"
      },
      "ConversationsListOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ConversationOut"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "ConversationsListOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageHistoryListOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "MessageHistoryListOut"
      },
      "MessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "sender",
          "content",
          "created_at"
        ],
        "title": "MessageOut"
      },
      "MessageStatsOut": {
        "properties": {
          "total_messages": {
            "type": "integer",
            "title": "Total Messages"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "by_character": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "By Character"
          },
          "by_day": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "By Day"
          }
        },
        "type": "object",
        "required": [
          "total_messages",
          "total_tokens",
          "total_cost"
        ],
        "title": "MessageStatsOut"
      },
...
      "OnebotMessageEntryOut": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "conversation_id",
          "character_id",
          "user_id",
          "sender",
          "content",
          "created_at"
        ],
        "title": "OnebotMessageEntryOut"
      },
      "OnebotMessagesListOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/OnebotMessageEntryOut"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "title": "OnebotMessagesListOut"
      },
...
      "SendMessageDataOut": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "content"
        ],
        "title": "SendMessageDataOut"
      },
      "SendMessageOut": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SendMessageDataOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SendMessageOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}